Renamed all folders under /src/compenents to use lowercase, merging folders with the same name but different capitalization.
