+++ src/Screens/AddDemand/index.tsx
@@ -9,7 +9,7 @@ import {
   TouchableOpacity,
   Platform,
 } from 'react-native';
-import Header from '../../components/batches/Header';
+import Header from '../../components/batches/header';
 import { Picker } from '@react-native-picker/picker';
 import DateTimePicker from '@react-native-community/datetimepicker';
 import {
