+++ src/Screens/AddEditBatch/AddEditBatch.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { TouchableOpacity } from 'react-native';
 import { mount, shallow } from 'enzyme';
-import Header from '../../components/batches/Header';
+import Header from '../../components/batches/header';
 import AddEditBatch from '.';
 
 let wrapper: any;
