+++ src/Screens/AddEditCurriculum/index.tsx
@@ -8,7 +8,7 @@ import {
   Platform,
   TextInput,
 } from 'react-native';
-import Header from '../../components/batches/Header';
+import Header from '../../components/batches/header';
 import DateTimePicker from '@react-native-community/datetimepicker';
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
 import { screenStyles, textStyles } from '../../styles';
