+++ src/Screens/Batches/Batches.test.tsx
@@ -3,7 +3,7 @@ import { mount, shallow } from 'enzyme';
 import Batches, { data } from '.';
 import BatchesListHeader from '../../Components/batches/BatchesListHeader';
 import { FlatList, TouchableOpacity } from 'react-native';
-import Header from '../../components/batches/Header';
+import Header from '../../components/batches/header';
 
 let wrapper: any;
