+++ src/Screens/Clients/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { SafeAreaView, FlatList } from 'react-native';
 import { screenStyles } from '../../styles';
-import ClientsListHeader from '../../components/clients/ClientsListHeader';
-import ClientsListItem from '../../components/clients/ClientsListItem';
+import ClientsListHeader from '../../components/clients/clients-list-header';
+import ClientsListItem from '../../components/clients/clients-list-item';
 
 /** Mock Data - PreRedux */
 export const data = ['Cognizant', 'Revature', 'Matts BBQ'];
