+++ src/Screens/ViewBatch/index.tsx
@@ -12,7 +12,7 @@ import { ProgressChart } from 'react-native-chart-kit';
 import { useNavigation } from '@react-navigation/native';
 import { RootStackParamList } from '../../types';
 import { StackNavigationProp } from '@react-navigation/stack';
-import Header from '../../components/batches/Header';
+import Header from '../../components/batches/header';
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
 import {
   badgesStyles,
