+++ src/Screens/ViewClient/index.tsx
@@ -10,11 +10,11 @@ import {
 import { useNavigation } from '@react-navigation/native';
 import { RootStackParamList } from '../../types';
 import { StackNavigationProp } from '@react-navigation/stack';
-import Header from '../../components/batches/Header';
+import Header from '../../components/batches/header';
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
 import { screenStyles, textStyles, buttonStyles, colors } from '../../styles';
 import { FlatList } from 'react-native-gesture-handler';
-import DemandsListItem from '../../components/demands/DemandsListItem';
+import DemandsListItem from '../../components/demands/demands-list-item';
 
 interface PropsI {
   route: {
