+++ src/components/curricula/curricula-list-header/index.tsx
