+++ src/components/curricula/expandable-list/index.tsx
